Simple version bumps
I'm working to update `haskell-src-exts` in the Gentoo Haskell packages, and this is one of the packages pinning it to a lower version.  I've tested on my system, and the upper bound there seems to be safe to bump to `< 1.24`; everything up to 1.23 builds with successful tests.  At the same time, `QuickCheck < 2.14` also seems safe.  I'll create an issue for `base` since there's a bit of fixing for that one that's not quite compatible.